Move RecordBatch::ToTensor code to tensor.cc
